requirements.txt
tiktoken==0.5.1

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3

# AI and ML
openai==1.3.5
faiss-cpu==1.7.4
numpy==1.25.2
scikit-learn==1.3.2

# Utilities
requests==2.31.0
aiofiles==23.2.1